Moving multisignature based functions to accounts
